requirements.txt
Flask==2.0.2
pdf2image==1.16.0
pytesseract==0.3.8
python-docx==0.8.11
Pillow==8.4.0
tqdm==4.62.3